Splitting out Census and Search
